support for folder sync via rsync on windows.
rsync interprets paths with colons as remote locations.
This leads to "The source and destination cannot both be remote." erros when using local paths like c:\vagrant for the hostpath.
This commits converts the hostpath to /cygdrive/... notation on windows in general.
I confirmed that this works with the cwRsync Installer (https://www.itefix.no/i2/cwrsync).