Retry the initial ssh to an ec2 instance
This is because we're still running into the same issue as described at
https://github.com/mitchellh/vagrant-aws/issues/74#issuecomment-54758045,
more than 2 years after it was intially described.

I'm not familiar enough with ruby to know how to discover/narrow down
the class of the exception - it doesn't appear to be part of the stack
trace - so I'm rescuing very broadly, and retrying.  Since there's a
retry limit, in the case where ssh is really unreachable this should
still terminate.

I'd perfer to only catch the network unreachable error.