Fix entry about "security_groups" in Readme.md
The entry on security groups, as written, does not fully make sense because it contrasts "VPC" and "nondefault VPC." But a "nodefault VPC" is still a "VPC"; in other words, the two categories actually overlap. The text should refer to a "default VPC" versus a "nondefault VPC"—categories that are mutually exclusive and exhaustive. 

While this may seem trivial, it confused me when I was trying to figure out whether to use Group Name or Group ID in my Vagrantfile. I have made some other minor changes that I think make the language even clearer.