Require vagrant 1.2 to 2.2.15
This is provided because 2.2.16 has a bug where it cannot be used with RSA
SHA1 ssh keys which makes it impossible to use with vagrant-aws as well as
any other platform that only supports RSA SHA1 keys. There's been no
activity from the vagrant maintainers to fix this for some time.

https://github.com/hashicorp/vagrant/issues/12344